DMP-647 Refactor default handler for Event API
Fixed code coverage issues by moving POJO to model
